Fix cache bug, refactored, moved cache key.
